refactor: relocate registration info (API)
