+++ src/index.ts
@@ -5,6 +5,7 @@ export { action } from "./actions/decorators";
 export { SingletonAction } from "./actions/singleton-action";
 export { DeviceType } from "./api/device";
 export * from "./api/layout";
+export { RegistrationInfo } from "./api/registration";
 export { Target } from "./api/target";
 export { EventEmitter, EventsOf } from "./common/event-emitter";
 export * from "./events";
